+++ backend/src/main/java/com/passtival/backend/domain/matching/controller/MatchingController.java
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.passtival.backend.domain.matching.model.response.MatchingResponse;
 import com.passtival.backend.domain.matching.service.MatchingService;
-import com.passtival.backend.global.auth.model.CustomMemberDetails;
+import com.passtival.backend.global.security.model.CustomMemberDetails;
 import com.passtival.backend.global.common.BaseResponse;
 
 import io.swagger.v3.oas.annotations.Operation;
