+++ backend/src/main/java/com/passtival/backend/domain/matching/controller/MemberController.java
@@ -10,7 +10,7 @@ import org.springframework.web.bind.annotation.RestController;
 import com.passtival.backend.domain.matching.model.request.MemberPatchRequest;
 import com.passtival.backend.domain.matching.model.response.MemberResponse;
 import com.passtival.backend.domain.matching.service.MemberService;
-import com.passtival.backend.global.auth.model.CustomMemberDetails;
+import com.passtival.backend.global.security.model.CustomMemberDetails;
 import com.passtival.backend.global.common.BaseResponse;
 
 import io.swagger.v3.oas.annotations.Operation;
