+++ backend/src/main/java/com/passtival/backend/global/security/controller/TokenController.java
@@ -1,13 +1,13 @@
-package com.passtival.backend.global.auth.controller;
+package com.passtival.backend.global.security.controller;
 
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.passtival.backend.global.auth.model.token.RefreshTokenRequest;
-import com.passtival.backend.global.auth.model.token.TokenResponse;
-import com.passtival.backend.global.auth.service.AuthService;
+import com.passtival.backend.global.security.model.token.RefreshTokenRequest;
+import com.passtival.backend.global.security.model.token.TokenResponse;
+import com.passtival.backend.global.security.service.TokenService;
 import com.passtival.backend.global.common.BaseResponse;
 import com.passtival.backend.global.exception.BaseException;
 
@@ -22,14 +22,14 @@ import lombok.RequiredArgsConstructor;
 @RequestMapping("/api/auth")
 @RequiredArgsConstructor
 @Tag(name = "인증/인가 관련 API", description = "토큰 갱신")
-public class AuthController {
+public class TokenController {
 
-	private final AuthService authService;
+	private final TokenService tokenService;
 
 	@PostMapping("/refresh")
 	public BaseResponse<TokenResponse> refreshToken(@Valid @RequestBody RefreshTokenRequest request) throws
 		BaseException {
-		TokenResponse tokenResponse = authService.refreshAccessToken(request);
+		TokenResponse tokenResponse = tokenService.refreshAccessToken(request);
 		return BaseResponse.success(tokenResponse);
 	}
 }
