+++ backend/src/main/java/com/passtival/backend/global/security/util/ResponseUtil.java
@@ -1,4 +1,4 @@
-package com.passtival.backend.global.auth.util;
+package com.passtival.backend.global.security.util;
 
 import java.io.IOException;
